Fix version transform for atomic
https://coveord.atlassian.net/browse/KIT-4606
